+++ packages/core/src/hooks/useConfig/examples/DummyComponent.tsx
@@ -1,0 +1,5 @@
+import { HassConfig } from "home-assistant-js-websocket";
+
+export function DummyComponentReturn(props: Partial<HassConfig>) {
+  return props;
+}