+++ packages/core/src/hooks/useHistory/examples/DummyComponent.tsx
@@ -1,0 +1,5 @@
+import { HistoryOptions } from '@hakit/core';
+
+export function DummyComponentHistoryOptions(props: Partial<HistoryOptions>) {
+  return props;
+}