+++ packages/core/src/hooks/useLightColor/examples/DummyComponent.tsx
@@ -1,0 +1,5 @@
+import { UseLightColorReturn } from "@hakit/core";
+
+export function DummyComponentReturn(props: Partial<UseLightColorReturn | null>) {
+  return props;
+}