+++ packages/core/src/hooks/useLowDevices/examples/DummyComponent.tsx
@@ -1,0 +1,5 @@
+import { LowDevicesOptions } from '@hakit/core';
+
+export function DummyComponentLowDevicesProps(props: Partial<LowDevicesOptions>) {
+  return props;
+}