+++ packages/core/src/hooks/useLowDevices/index.ts
@@ -4,19 +4,19 @@ import { HassEntity } from "home-assistant-js-websocket";
 import { useShallow } from "zustand/shallow";
 
 export interface LowDevicesOptions {
-  /** the minimum battery percentage threshold */
+  /** The minimum battery percentage to retrieve @default 0 */
   min?: number;
-  /** the maximum battery percentage threshold */
+  /** The maximum battery percentage to retrieve @default 20 */
   max?: number;
-  /** the list of entities to ignore */
+  /** If there's entities returning in the results, that you want to exclude, you can provide a partial entity_id match to exclude it @default [] */
   blacklist?: string[];
-  /** the list of entities to only include */
+  /** If there's entities returning in the results, but you only want certain entities, provide a partial entity_id match to include them @default [] */
   whitelist?: string[];
 }
 
 const isBattery = (e: HassEntity) => e.attributes.unit_of_measurement === "%" && e.attributes.device_class === "battery";
 
-export const useLowDevices = ({ blacklist = [], whitelist = [], min = 0, max = 20 }: LowDevicesOptions = {}) => {
+export const useLowDevices = ({ blacklist = [], whitelist = [], min = 0, max = 20 }: LowDevicesOptions = {}): HassEntity[] => {
   const batteries = useStore(
     useShallow((state) => {
       return Object.values(state.entities).filter(isBattery);