+++ packages/core/src/hooks/useTemplate/examples/DummyComponent.tsx
@@ -1,0 +1,5 @@
+
+import { TemplateParams } from '@hakit/core';
+export function DummyComponentOptions(props: Partial<TemplateParams>) {
+  return props;
+}