+++ packages/core/src/hooks/useWeather/examples/DummyComponent.tsx
@@ -1,0 +1,9 @@
+import { WeatherEntityWithForecast, UseWeatherOptions } from "@hakit/core";
+
+export function DummyComponentOptions(props: Partial<UseWeatherOptions>) {
+  return props;
+}
+
+export function DummyComponent(props: Partial<WeatherEntityWithForecast>) {
+  return props;
+}