+++ packages/core/src/hooks/useWeather/helpers.ts
@@ -71,16 +71,16 @@ const getLegacyForecast = (
   return undefined;
 };
 
+export type Forecast = {
+  forecast: ForecastAttribute[];
+  type: "daily" | "hourly" | "twice_daily";
+} | undefined
+
 export const getForecast = (
   weather_attributes: WeatherEntity["attributes"],
   forecast_event: ForecastEvent | null,
   forecast_type?: ForecastType | undefined,
-):
-  | {
-      forecast: ForecastAttribute[];
-      type: "daily" | "hourly" | "twice_daily";
-    }
-  | undefined => {
+): Forecast => {
   if (forecast_type === undefined) {
     if (forecast_event?.type !== undefined && forecast_event?.forecast && forecast_event?.forecast?.length > 2) {
       return { forecast: forecast_event.forecast, type: forecast_event?.type };